+++ packages/apps/terminal/src/core/commands/reboot.ts
@@ -1,0 +1,11 @@
+import { reloadViewport } from "@prozilla-os/core";
+import { Command } from "../command";
+
+export const reboot = new Command()
+	.setManual({
+		purpose: "Reboot the system"
+	})
+	.setExecute(function() {
+		reloadViewport();
+		return { blank: true };
+	});