+++ packages/core/src/components/modals/index.ts
@@ -5,3 +5,4 @@ export { Share } from "./share/Share";
 export { ModalsView } from "./ModalsView";
 
 export type { ModalProps } from "./ModalView";
+export { FileSelectorMode } from "@prozilla-os/file-explorer";