+++ packages/core/src/constants/index.ts
@@ -1,3 +1,4 @@
 export { DIALOG_CONTENT_TYPES } from "./modals.const";
 export { CODE_EXTENSIONS } from "./virtualDrive.const";
 export { ANSI } from "./utils.const";
+export { THEMES } from "./themes.const";