+++ packages/prozilla-os/src/index.ts
@@ -1,2 +1,4 @@
 export * from "@prozilla-os/core";
 export * from "@prozilla-os/file-explorer";
+export * from "@prozilla-os/terminal";
+export * from "@prozilla-os/text-editor";