+++ src/config/branding.config.ts
@@ -1,5 +1,3 @@
-import { ANSI } from "prozilla-os";
-
 export const NAME = "ProzillaOS";
 export const TAG_LINE = "Web-based Operating System";
 
@@ -15,24 +13,6 @@ export const ASCII_LOGO = `
 :==+----#%+-+%#-##%*+----:.
   .=----#%+-+%#-*+-%#+---:.
    ==----*###*--*###*----. 
-  ==+-------------------:. 
-  ...::---------------:.   
-       .::---------::..    
-          ....::...        `;
-
-export const ANSI_LOGO_COLOR = ANSI.fg.cyan;
-export const ANSI_ASCII_LOGO = `
-              :.           
-             -==.          
-           .=====:         
-   ---::..:=======-.       
-   :===+=----------::..    
-    =+=---------------:..  
-    --------------------:. 
-.:-+=----${ANSI.fg.white}*###*${ANSI_LOGO_COLOR}--${ANSI.fg.white}*####=${ANSI_LOGO_COLOR}---. 
-:==+----${ANSI.fg.white}#%+${ANSI_LOGO_COLOR}-${ANSI.fg.white}+%#${ANSI_LOGO_COLOR}-${ANSI.fg.white}##%*+${ANSI_LOGO_COLOR}----:.
-  .=----${ANSI.fg.white}#%+${ANSI_LOGO_COLOR}-${ANSI.fg.white}+%#${ANSI_LOGO_COLOR}-${ANSI.fg.white}*+${ANSI_LOGO_COLOR}-${ANSI.fg.white}%#+${ANSI_LOGO_COLOR}---:.
-   ==----${ANSI.fg.white}*###*${ANSI_LOGO_COLOR}--${ANSI.fg.white}*###*${ANSI_LOGO_COLOR}----. 
   ==+-------------------:. 
   ...::---------------:.   
        .::---------::..