+++ packages/prozilla-os/CHANGELOG.md
@@ -1,5 +1,16 @@
 # prozilla-os
 
+## 1.0.11
+
+### Patch Changes
+
+- Added terminal and text editor app + adjusted links in app packages
+- Updated dependencies
+  - @prozilla-os/file-explorer@1.0.8
+  - @prozilla-os/core@1.0.11
+  - @prozilla-os/terminal@1.0.1
+  - @prozilla-os/text-editor@1.0.1
+
 ## 1.0.10
 
 ### Patch Changes